+++ axum/src/macros.rs
@@ -77,6 +77,14 @@ macro_rules! trace {
 
 #[cfg(feature = "tracing")]
 #[allow(unused_macros)]
+macro_rules! info {
+    ($($tt:tt)*) => {
+        tracing::info!($($tt)*)
+    }
+}
+
+#[cfg(feature = "tracing")]
+#[allow(unused_macros)]
 macro_rules! error {
     ($($tt:tt)*) => {
         tracing::error!($($tt)*)
@@ -86,6 +94,12 @@ macro_rules! error {
 #[cfg(not(feature = "tracing"))]
 #[allow(unused_macros)]
 macro_rules! trace {
+    ($($tt:tt)*) => {};
+}
+
+#[cfg(not(feature = "tracing"))]
+#[allow(unused_macros)]
+macro_rules! info {
     ($($tt:tt)*) => {};
 }